feat: add uncaught and unhandled errors handlers
